butler/livekit-agent/requirements.txt
livekit-agents~=1.3
livekit-plugins-silero~=1.3
livekit-plugins-openai~=1.3
livekit-plugins-groq~=1.3
aiohttp
pydantic-settings
python-dotenv
